add script to compile and run c++ code

+++ src/cpp_api.cpp
@@ -6,7 +6,7 @@
 
 namespace std {
 
-int cpp_api(int a, int b) {
+double cpp_api(double a, double b) {
 	return a + b;
 }
